fix media asset page tests
